Beta-v1
The game is all ready to play! It's my first game, and it's simple for now, but the goal is to add many tweaks here and there, to make this simple game be actually fun.
List of things that are new:
- Background color: You can change the bg color to the three main ones (purple, blue and red). Soon I will add the option for you to choose a color of your choice.
- Sounds: There are some sounds played when you click on things. I.e. when you eat a food, or click a button. I also implemented the option for you to change the volume of the sound effects in the settings button.
That's it for now! But I'm already planning to add many things to this game, and I hope it turns out good :)